# LearningCenterPlatform/bin/Release/net9.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ACME.LearningCenterPlatform.API",
    "description": "ACME Learning Center Platform API",
    "termsOfService": "https://acme-learning.com/tos",
    "contact": {
      "name": "ACME Studios",
      "email": "contact@acme.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/tutorials/{tutorialId}": {
      "get": {
        "tags": [
          "Tutorials"
        ],
        "summary": "Get a tutorial by its id",
        "description": "Get a tutorial by its id",
        "operationId": "GetTutorialById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "tutorialId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "The tutorial was found",
            "schema": {
              "$ref": "#/definitions/TutorialResource"
            }
          },
          "404": {
            "description": "The tutorial was not found"
          }
        }
      }
    },
    "/api/v1/tutorials": {
      "post": {
        "tags": [
          "Tutorials"
        ],
        "summary": "Create a tutorial",
        "description": "Create a tutorial",
        "operationId": "CreateTutorial",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateTutorialResource"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "The tutorial was created",
            "schema": {
              "$ref": "#/definitions/TutorialResource"
            }
          },
          "400": {
            "description": "The tutorial was not created"
          }
        }
      },
      "put": {
        "tags": [
          "Tutorials"
        ],
        "summary": "Update a tutorial",
        "description": "Update a tutorial",
        "operationId": "UpdateTutorial",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateTutorialResource"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "The tutorial was updated",
            "schema": {
              "$ref": "#/definitions/CategoryResource"
            }
          },
          "400": {
            "description": "The tutorial could not be updated"
          }
        }
      },
      "delete": {
        "tags": [
          "Tutorials"
        ],
        "summary": "Delete a Tutorial",
        "description": "Detele a Tutorial",
        "operationId": "DeleteTutorial",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DeleteTutorialResource"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "The Tutorial was deleted",
            "schema": {
              "$ref": "#/definitions/CategoryResource"
            }
          },
          "400": {
            "description": "The Tutorial could not be Deleted"
          }
        }
      },
      "get": {
        "tags": [
          "Tutorials"
        ],
        "summary": "Get all tutorials",
        "description": "Get all tutorials",
        "operationId": "GetAllTutorials",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "The tutorials were found",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TutorialResource"
              }
            }
          }
        }
      }
    },
    "/api/v1/tutorials/{tutorialId}/videos": {
      "post": {
        "tags": [
          "Tutorials"
        ],
        "summary": "Add a video to a tutorial",
        "description": "Add a video to a tutorial",
        "operationId": "AddVideoToTutorial",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "tutorialId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AddVideoAssetToTutorialResource"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "The video was added to the tutorial",
            "schema": {
              "$ref": "#/definitions/TutorialResource"
            }
          },
          "400": {
            "description": "The video was not added to the tutorial"
          }
        }
      }
    },
...
  },
  "definitions": {
    "AddVideoAssetToTutorialResource": {
      "type": "object",
      "properties": {
        "videoUrl": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "CategoryResource": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "CreateTutorialResource": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "summary": {
          "type": "string"
        },
        "categoryId": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
...
    "DeleteTutorialResource": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
...
    "TutorialResource": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "summary": {
          "type": "string"
        },
        "category": {
          "$ref": "#/definitions/CategoryResource"
        },
        "status": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "UpdateTutorialResource": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "summary": {
          "type": "string"
        },
        "categoryId": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
...
  },
  "securityDefinitions": {
    "Bearer": { }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}